2917 - Surface `message` when available for errors from `client`
`internalClient` was swallowing up an `error` and just using `statusText`.

I modified the error message to continue to do that, but also check the `res` for a `message` key and use that, as well.

This should result in more useful error messages (specifically, a `404` not only responds with `Not Found`, but also what document the server failed to find).